User major axis for comparison in assignmentPlausibilityTester.sizeDifferenceIsPlausible.

+++ src/main/java/com/jug/lp/AssignmentPlausibilityTester.java
@@ -23,7 +23,7 @@ public class AssignmentPlausibilityTester {
      * @param totalTargetComponentSize
      * @return
      */
-    public boolean sizeDifferenceIsPlausible(long sourceComponentSize, long totalTargetComponentSize) {
+    public boolean sizeDifferenceIsPlausible(double sourceComponentSize, double totalTargetComponentSize) {
         double foldChange = ((double) totalTargetComponentSize) / ((double) sourceComponentSize);
         boolean res = foldChange <= sizeDifferenceThreshold;
         return res;
